notes and player init
