+++ source/OTRServer/routes/player.js
@@ -1,0 +1,10 @@
+var express = require('express');
+var path = require('path')
+var router = express.Router()
+
+
+router.get("/", (req, res) => {
+    res.sendFile(path.join(__dirname, '../views/player.html'))
+})
+
+module.exports = router
