+++ source/OTRServer/views/admin.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>OnTheRun Admin Panel</title>
+</head>
+<body>
+    
+</body>
+</html>
