+++ source/OTRServer/views/player.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>OnTheRun</title>
+    <link rel="stylesheet" type="text/css" href="./css/player.css">
+    <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:ital,wght@0,400..700;1,400..700&display=swap" rel="stylesheet">
+    <script src="https://cdn.socket.io/4.7.5/socket.io.min.js" integrity="sha384-2huaZvOR9iDzHqslqwpR87isEmrfxqyWOF7hr7BY6KG0+hVKLoEXMPUJw3ynWuhO" crossorigin="anonymous"></script>
+    <script src="https://cdn.socket.io/4.7.5/socket.io.min.js" integrity="sha384-2huaZvOR9iDzHqslqwpR87isEmrfxqyWOF7hr7BY6KG0+hVKLoEXMPUJw3ynWuhO" crossorigin="anonymous"></script>
+</head>
+<body>
+    <div id="wrapper">
+        <div id="login-wrapper">
+            <img id="login-hero" src="./images/PlayerHero.png">
+            <div id="login-username">
+                <input id="login-username-input" type="text" placeholder="Your Name" >
+            </div>
+            <div id="login-submit">
+                <div id="login-button">Join</div>
+            </div>
+        </div>
+    </div>
+    <script src="./player.js"></script>
+</body>
+</html>
