+++ source/notes.md
@@ -1,0 +1,20 @@
+# Notes
+
+## Running
+
+### Start Server 
+```bash
+    cd source/OTRServer
+    node .
+```
+
+### Public Routes  
+Player:
+```
+localhost:3000
+```
+Display: 
+```
+localhost:3000/display
+```
+
